requirements.txt
sw-django-utils